feat: start to simplify multi-target structure
