+++ Makefile
@@ -1,8 +1,8 @@
 start:
-	deno run --allow-read --allow-write --allow-env --allow-run source/deno-cli/main.ts
+	deno run --allow-read --allow-write --allow-env --allow-run source/main.ts
 
 install:
-	deno install --allow-read --allow-write --allow-env --allow-run --name=jd source/deno-cli/main.ts
+	deno install --allow-read --allow-write --allow-env --allow-run --name=jd source/main.ts
 
 uninstall:
 	deno uninstall jd
