--- source/core/types.ts
@@ -1,7 +1,0 @@
-export interface Command {
-  name: string;
-  alias?: string[];
-  description?: string;
-  helpText?: string;
-  (...args: any[]): any;
-}
