+++ source/deps.ts
@@ -1,0 +1,30 @@
+import {
+  copySync,
+  ensureDir,
+  exists,
+  walk,
+} from "https://deno.land/std/fs/mod.ts";
+import {
+  basename,
+  dirname,
+  fromFileUrl,
+  join,
+} from "https://deno.land/std/path/mod.ts";
+import { bold, red } from "https://deno.land/std/fmt/colors.ts";
+import { parse } from "https://deno.land/std/flags/mod.ts";
+import { config } from "https://deno.land/std/dotenv/mod.ts";
+
+export {
+  basename,
+  bold,
+  config,
+  copySync,
+  dirname,
+  ensureDir,
+  exists,
+  fromFileUrl,
+  join,
+  parse,
+  red,
+  walk,
+};
